+++ CHANGELOG.md
@@ -9,7 +9,8 @@ This page shows a detailed overview of the changes between versions without the 
 
 ## **WORK IN PROGRESS**
 
-- Speed up attribute data conversion for migration and websocket
+- Speed up attribute data conversion for migration and websocket communication
+- Enhancement: Update matter.js to 0.17.4
 
 ## 1.1.6 (2026-06-30)